Beginning shell

+++ include/utils.h
@@ -3,7 +3,7 @@
 
 #include <stdint.h>
 #include <stddef.h>
-int strlen(const char* str);
+size_t strlen(const char* str);
 void int_to_string(int num, char* buffer);
 uint16_t low_16(uint32_t addr);
 uint16_t high_16(uint32_t addr);
